Refactor: 로그인 페이지 통합
Refs: #7
